+++ project/Build.scala
@@ -13,19 +13,25 @@ object SummingbirdBuild extends Build {
       case x => x
     }
 
+  def specs2Import(scalaVersion: String) = scalaVersion match {
+      case version if version startsWith "2.9" => "org.specs2" %% "specs2" % "1.12.4.1" % "test"
+      case version if version startsWith "2.10" => "org.specs2" %% "specs2" % "1.13" % "test"
+  }
+
   val sharedSettings = Project.defaultSettings ++ Seq(
     organization := "com.twitter",
-    version := "0.2.4",
+    version := "0.2.5",
     scalaVersion := "2.9.3",
     crossScalaVersions := Seq("2.9.3", "2.10.0"),
     libraryDependencies ++= Seq(
       "org.slf4j" % "slf4j-api" % slf4jVersion,
       "org.scalacheck" %% "scalacheck" % "1.10.0" % "test",
-      "org.scala-tools.testing" %% "specs" % "1.6.9" % "test",
       // These satisify's scaldings log4j needs when in test mode
       "log4j" % "log4j" % "1.2.16" % "test",
       "org.slf4j" % "slf4j-log4j12" % slf4jVersion % "test"
     ),
+
+    libraryDependencies <+= scalaVersion(specs2Import(_)),
 
     resolvers ++= Seq(
       Opts.resolver.sonatypeSnapshots,
@@ -109,14 +115,15 @@ object SummingbirdBuild extends Build {
     summingbirdStorm,
     summingbirdScalding,
     summingbirdBuilder,
+    summingbirdChill,
     summingbirdExample
   )
 
   val dfsDatastoresVersion = "1.3.4"
   val bijectionVersion = "0.5.4"
   val algebirdVersion = "0.3.0"
-  val scaldingVersion = "0.9.0rc1"
-  val storehausVersion = "0.6.0"
+  val scaldingVersion = "0.9.0rc4"
+  val storehausVersion = "0.7.2"
   val utilVersion = "6.3.8"
   val chillVersion = "0.3.3"
   val tormentaVersion = "0.5.3"
@@ -132,7 +139,7 @@ object SummingbirdBuild extends Build {
   def youngestForwardCompatible(subProj: String) =
     Some(subProj)
       .filterNot(unreleasedModules.contains(_))
-      .map { s => "com.twitter" % ("summingbird-" + s + "_2.9.3") % "0.1.0" }
+      .map { s => "com.twitter" % ("summingbird-" + s + "_2.9.3") % "0.2.4" }
 
   def module(name: String) = {
     val id = "summingbird-%s".format(name)
@@ -147,6 +154,15 @@ object SummingbirdBuild extends Build {
       "com.twitter" %% "algebird-core" % algebirdVersion,
       "com.twitter" %% "bijection-core" % bijectionVersion
     )
+  )
+
+  lazy val summingbirdChill = module("chill").settings(
+    libraryDependencies ++= Seq(
+      "com.twitter" %% "chill" % chillVersion
+    )
+  ).dependsOn(
+      summingbirdCore,
+      summingbirdBatch
   )
 
   lazy val summingbirdClient = module("client").settings(
@@ -194,6 +210,7 @@ object SummingbirdBuild extends Build {
   ).dependsOn(
     summingbirdCore % "test->test;compile->compile",
     summingbirdOnline,
+    summingbirdChill,
     summingbirdBatch
   )
 
@@ -214,6 +231,7 @@ object SummingbirdBuild extends Build {
     )
   ).dependsOn(
     summingbirdCore % "test->test;compile->compile",
+    summingbirdChill,
     summingbirdBatch
   )