+++ summingbird-scalding/src/main/scala/com/twitter/summingbird/scalding/ScaldingPlatform.scala
@@ -25,6 +25,8 @@ import com.twitter.scalding.{ Tool => STool, Source => SSource, TimePathedSource
 import com.twitter.summingbird._
 import com.twitter.summingbird.scalding.option.{ FlatMapShards, Reducers }
 import com.twitter.summingbird.batch._
+import com.twitter.chill.IKryoRegistrar
+import com.twitter.summingbird.chill._
 import com.twitter.summingbird.option._
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.Path
@@ -42,6 +44,11 @@ import org.slf4j.LoggerFactory
 
 object Scalding {
   @transient private val logger = LoggerFactory.getLogger(classOf[Scalding])
+
+
+  def apply(jobName: String, options: Map[String, Options] = Map.empty) = {
+    new Scalding(jobName, options, identity, List())
+  }
 
   implicit val dateRangeInjection: Injection[DateRange, Interval[Time]] = Injection.build {
     (dr: DateRange) => {
@@ -225,10 +232,10 @@ object Scalding {
 
   private def getOrElse[T: Manifest](options: Map[String, Options], idOpt: Option[String], default: T): T =
     (for {
-      id <- idOpt
+      id <- idOpt.map(List(_, "DEFAULT")).getOrElse(List("DEFAULT"))
       innerOpts <- options.get(id)
       option <- innerOpts.get[T]
-    } yield option).getOrElse(default)
+    } yield option).headOption.getOrElse(default)
 
   @annotation.tailrec
   private def getFirst[T: Manifest](options: Map[String, Options], names: List[String]): Option[T] =
@@ -494,7 +501,10 @@ case class WriteStepsDot(filename: String)
 
 class Scalding(
   jobName: String,
-  @transient options: Map[String, Options] = Map.empty)
+  @transient options: Map[String, Options],
+  @transient transformConfig: SummingbirdConfig => SummingbirdConfig,
+  @transient passedRegistrars: List[IKryoRegistrar]
+  )
     extends Platform[Scalding] {
 
   type Source[T] = PipeFactory[T]
@@ -512,12 +522,30 @@ class Scalding(
     classOf[com.twitter.chill.hadoop.KryoSerialization]
   )
 
-  private def setIoSerializations(m: Mode): Unit =
-    m match {
-      case Hdfs(_, conf) =>
-        conf.set("io.serializations", ioSerializations.map { _.getName }.mkString(","))
-      case _ => ()
-    }
+  def withRegistrars(newRegs: List[IKryoRegistrar]) =
+    new Scalding(jobName, options, transformConfig, newRegs ++ passedRegistrars)
+
+  def withConfigUpdater(fn: SummingbirdConfig => SummingbirdConfig) =
+    new Scalding(jobName, options, transformConfig.andThen(fn), passedRegistrars)
+
+  def updateConfig(conf: Configuration) {
+    val scaldingConfig = SBChillRegistrar(ScaldingConfig(conf), passedRegistrars)
+    Scalding.logger.debug("Serialization config changes:")
+    Scalding.logger.debug("Removes: {}", scaldingConfig.removes)
+    Scalding.logger.debug("Updates: {}", scaldingConfig.updates)
+
+    val transformedConfig = transformConfig(scaldingConfig)
+
+    Scalding.logger.debug("User+Serialization config changes:")
+    Scalding.logger.debug("Removes: {}", transformedConfig.removes)
+    Scalding.logger.debug("Updates: {}", transformedConfig.updates)
+
+    transformedConfig.removes.foreach(conf.set(_, null))
+    transformedConfig.updates.foreach(kv => conf.set(kv._1, kv._2.toString))
+  }
+
+  private def setIoSerializations(c: Configuration): Unit =
+      c.set("io.serializations", ioSerializations.map { _.getName }.mkString(","))
 
   // This is a side-effect-free computation that is called by run
   def toFlow(timeSpan: Interval[Time], mode: Mode, pf: PipeFactory[_]): Try[(Interval[Time], Flow[_])] = {
@@ -544,7 +572,14 @@ class Scalding(
     mode: Mode,
     pf: PipeFactory[Any]): WaitingState[Interval[Timestamp]] = {
 
-    setIoSerializations(mode)
+    mode match {
+      case Hdfs(_, conf) =>
+        updateConfig(conf)
+        setIoSerializations(conf)
+      case _ =>
+    }
+
+
 
     val prepareState = state.begin
     val timeSpan = prepareState.requested.mapNonDecreasing(_.milliSinceEpoch)